refactor(memory): remove MemoryProvider abstraction; file is the only backend
Delete src/memory/interface.ts, factory.ts, openviking.ts, mem0.ts, and
test/test-openviking.ts. Rename FileMemoryProvider to MemoryStore; move
Episode/EpisodeMeta/Skill type definitions inline.

Rewire index.ts to instantiate MemoryStore directly. Rename memoryProvider
field to memoryStore (and setMemoryProvider -> setMemoryStore) in channel.ts
and tools.ts.

Remove MEMORY_PROVIDER, OPENVIKING_URL, OPENVIKING_API_KEY, MEM0_URL,
MEM0_API_KEY config fields and .env.example entries.

Users on MEMORY_PROVIDER=file (the default): zero action required.
Users previously on MEMORY_PROVIDER=openviking: local-file memory is
preserved (OpenViking already used local files as the profile hot path);
only the Viking-side vector index is abandoned.

Precondition for the privacy redesign (#35). A pluggable abstraction made
every downstream interface change a three-file synchronization exercise,
and the Viking vector index raised a policy question (index private-tier
content?) with no good default for a one-operator plugin.
