+++ .env.example
@@ -9,15 +9,6 @@ LARK_APP_SECRET=
 # LARK_ENABLED_SKILLS=           # lark-cli skill whitelist for start.sh
 
 # === Optional — Memory ===
-# MEMORY_PROVIDER=file           # Memory backend: file | openviking | mem0
-# LARK_MIN_SEARCH_SCORE=0.3     # Minimum relevance score (vector backends only)
+# LARK_MIN_SEARCH_SCORE=0.3     # Minimum relevance score for episode search
 # LARK_MAX_SEARCH_RESULTS=2     # Max episodes per query
 # LARK_INACTIVITY_HOURS=3       # Auto-flush trigger threshold
-
-# === Optional — OpenViking ===
-# OPENVIKING_URL=http://localhost:1933
-# OPENVIKING_API_KEY=
-
-# === Optional — mem0 ===
-# MEM0_URL=
-# MEM0_API_KEY=
